fix(timers): adopt omp_get_wtime(); widen legacy 32-bit timer helpers
Two related timing bugs reported by an upstream PR
(OceanLedger team / Claude-anthropic):

- Bug 1 (integer overflow): 32-bit count / count_rate / count_max
  locals with nanosecond clocks roll over every ~2.15 s, producing
  silent negative elapsed times and the "Time in momentum: 335%"
  symptom. The recent sfincs_timers refactor already fixed this in
  the main timer machinery; the two legacy timer(t) helpers in
  sfincs_date.f90 and snapwave/snapwave_solver.f90 still had
  integer*4 locals and are widened here.

- Bug 2 (CPU time vs wall time): system_clock under nvfortran +
  OpenACC GPU offload can return CPU time, not wall time, making
  a 6-hour GPU run report as ~10 minutes and causing time-remaining
  estimates to reset every ~10%. sfincs_timers.f90 now uses
  omp_get_wtime() (real*8 wall seconds, OpenMP-guaranteed
  monotonic) instead of system_clock.

No behavior change on CPU builds (wall clocks match within jitter).
No public signature changes. Culvert baseline unchanged at
crosssection_discharge mean 3.6939 m^3/s.

Credit: OceanLedger team for the diagnosis.

+++ source/src/sfincs_date.f90
@@ -365,7 +365,7 @@ CONTAINS
    !
    subroutine timer(t)
    real*4,intent(out)               :: t
-   integer*4                        :: count,count_rate,count_max
+   integer*8                        :: count,count_rate,count_max
    call system_clock (count,count_rate,count_max)
    t = dble(count)/count_rate
    end subroutine timer
